+++ src/README.tpl.md
@@ -1,6 +1,6 @@
 # Jamf School API for Deno
 
-Write powerful Jamf School automations with little effort.
+Create Jamf School automations, quickly and easily.
 
 An **unofficial, idiomatic API wrapper for Jamf School**. It's designed to be easy for IT professionals with JavaScript experience to learn, and to be safe enough to trust in production.
 
@@ -48,6 +48,8 @@ deno run --allow-net=YOUR_SCHOOL.jamfcloud.com device_names.js
 ```
 
 There's also a lower-level API wrapper available. [Here's the documentation]($DOCS/api.ts).
+
+For some more complete (public domain!) examples, check out the [examples](./examples) directory.
 
 ## Changelog
