+++ src/examples/README.md
@@ -1,0 +1,5 @@
+# Examples
+
+The code in this directory is public domain. It can be freely copied and reused. See [UNLICENSE](./UNLICENSE) for the legal stuff.
+
+Note that no other code in this repository is public domain, that's all under the [MIT license](../LICENSE).
